Fix forward compilation error and add Makefile for ANNA program

+++ HideWordSolver/Lib/matrix_product.c
@@ -1,8 +1,8 @@
 #include <err.h>
 #include "lib.h"
 
-void matrix_product(size_t row1, size_t col1, int mat1[row1][col1],
-		size_t row2, size_t col2, int mat2[row2][col2], int res[row1][col2])
+void matrix_product(size_t row1, size_t col1, double mat1[row1][col1],
+		size_t row2, size_t col2, double mat2[row2][col2], double res[row1][col2])
 {
 	if(col1 == row2)
 	{
@@ -10,7 +10,7 @@ void matrix_product(size_t row1, size_t col1, int mat1[row1][col1],
 		{
 			for(size_t j = 0; j < col2; j++)
 			{
-				int add = 0;
+				double add = 0;
 				for(size_t k = 0; k < col1; k++)
 				{
 					add += mat1[i][k] * mat2[k][j];
